Refactor ConditionForm to per-mode strategy polymorphism
Replace the five state.mode conditional ladders with RelationalMode /
CensusMode / IdentityMode strategies behind a uniform interface;
ConditionForm becomes pure dispatch (1085 → 329 lines). Shared select/
pill helpers extracted to condition_form/dom_helpers.js. CensusMode's
internal region state is renamed position*→region* (wire unchanged;
fromNodeData/buildPayload translate). Behavior-preserving — full JS
suite green; no Ruby touched
